move helper route in session controller to test file
